[Add] class : Payment
