+++ src/main/java/lotto/model/InvalidLottoNumberException.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 public class InvalidLottoNumberException extends RuntimeException {
         public InvalidLottoNumberException() {
