+++ src/main/java/lotto/model/PaymentNotMultipleOfThousandException.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 public class PaymentNotMultipleOfThousandException extends RuntimeException {
         public PaymentNotMultipleOfThousandException() {
