+++ src/main/java/lotto/model/PaymentNotNaturalNumberException.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 public class PaymentNotNaturalNumberException extends RuntimeException {
         public PaymentNotNaturalNumberException() {
