+++ src/main/java/lotto/model/WebUILottoApplication.java
@@ -1,0 +1,22 @@
+package lotto.model;
+
+import spark.ModelAndView;
+import spark.template.handlebars.HandlebarsTemplateEngine;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static spark.Spark.get;
+
+public class WebUILottoApplication {
+        public static void main(String[] args) {
+                get("/", (req, res) -> {
+                        Map<String, Object> model = new HashMap<>();
+                        return render(model, "index.html");
+                });
+        }
+
+        private static String render(Map<String, Object> model, String templatePath) {
+                return new HandlebarsTemplateEngine().render(new ModelAndView(model, templatePath));
+        }
+}
