+++ src/test/java/lotto/model/LottoNumberTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import org.junit.jupiter.api.Test;
 
@@ -7,15 +7,15 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 public class LottoNumberTest {
         @Test
         void 로또_숫자_범위_검사_0() {
-                assertThrows(InvalidLottoNumberException.class, ()->{
-                    LottoNumber.create(0);
-                } );
+                assertThrows(InvalidLottoNumberException.class, () -> {
+                        LottoNumber.create(0);
+                });
         }
 
         @Test
         void 로또_숫자_범위_검사_46() {
-                assertThrows(InvalidLottoNumberException.class, ()->{
+                assertThrows(InvalidLottoNumberException.class, () -> {
                         LottoNumber.create(46);
-                } );
+                });
         }
 }
